netdev-offload-dpdk: Enable re-release contextes
A pending context item is marked so it won't be submitted again for
release. In case it is re-used before it was released, its refcount
increases and skipped in the release queue. If it should be released
again it will not as it wrongly mark as pending to be released. Fix it.

Issue: 2239936
Fixes: 29845d092fdc ("netdev-offload-dpdk: Delayed context release")
Change-Id: I41ab8c7f6b121eec3f09c208396f042ee2bc5460
Signed-off-by: Eli Britstein <elibr@mellanox.com>