netdev-offload-dpdk: Fix multi-zone with zero zone
The cited commit meant to remove the ct-state match of flows with
recirc_id=0, as not needed. It also removed the match of zero CT zone
of such flows. In case of a CT flow, it must have that match, so it's
wrong. There is no match on CT-zone anyway on flows with recirc_id=0, so
reverting this part.

Issue: 2262467
Change-Id: I77f0be1a825fcc1f0231e633dee388f91b45f3f9
Fixes: a1260dfcb929 ("netdev-offload-dpdk: Don't match zeros on CT properties zero recirc_id")
Signed-off-by: Eli Britstein <elibr@mellanox.com>